BootLintBear: Move to HTML section
The BootLintBear is now in the HTML folder.

Fixes https://github.com/coala-analyzer/coala-bears/issues/158
